+++ main.go
@@ -29,10 +29,11 @@ func main() {
 }
 
 type application struct {
-	SentryDSN    string `required:"true"  arg:"sentry-dsn"    env:"SENTRY_DSN"    usage:"SentryDSN"                             display:"length"`
-	SentryProxy  string `required:"false" arg:"sentry-proxy"  env:"SENTRY_PROXY"  usage:"Sentry Proxy"`
-	Listen       string `required:"true"  arg:"listen"        env:"LISTEN"        usage:"address to listen to"`
-	KafkaBrokers string `required:"true"  arg:"kafka-brokers" env:"KAFKA_BROKERS" usage:"Comma separated list of Kafka brokers"`
+	SentryDSN                 string `required:"true"  arg:"sentry-dsn"                   env:"SENTRY_DSN"                   usage:"SentryDSN"                                                               display:"length"`
+	SentryProxy               string `required:"false" arg:"sentry-proxy"                 env:"SENTRY_PROXY"                 usage:"Sentry Proxy"`
+	Listen                    string `required:"true"  arg:"listen"                       env:"LISTEN"                       usage:"address to listen to"`
+	KafkaBrokers              string `required:"true"  arg:"kafka-brokers"                env:"KAFKA_BROKERS"                usage:"Comma separated list of Kafka brokers"`
+	ErrorPreviewContentLength int    `required:"false" arg:"error-preview-content-length" env:"ERROR_PREVIEW_CONTENT_LENGTH" usage:"Maximum length in bytes for error message preview. Use -1 for unlimited"                  default:"100"`
 }
 
 func (a *application) Run(
@@ -69,7 +70,8 @@ func (a *application) createHTTPServer(
 		router.Path("/setloglevel/{level}").Handler(
 			log.NewSetLoglevelHandler(ctx, log.NewLogLevelSetter(2, 5*time.Minute)),
 		)
-		router.Path("/read").Handler(factory.CreateReadHandler(sentryClient, saramaClient))
+		router.Path("/read").
+			Handler(factory.CreateReadHandler(sentryClient, saramaClient, a.ErrorPreviewContentLength))
 
 		glog.V(2).Infof("starting http server listen on %s", a.Listen)
 		return libhttp.NewServer(